+++ Source/cmExportInstallFileGenerator.cxx
@@ -435,7 +435,7 @@ cmExportInstallFileGenerator::HandleMissingTarget(
 {
   const std::string name = dependee->GetName();
   std::vector<std::string> namespaces = this->FindNamespaces(mf, name);
-  int targetOccurrences = (int)namespaces.size();
+  size_t targetOccurrences = namespaces.size();
   if (targetOccurrences == 1)
     {
     std::string missingTarget = namespaces[0];
@@ -470,7 +470,7 @@ cmExportInstallFileGenerator
                                                  exportSet->GetTargetExports();
 
     bool containsTarget = false;
-    for(unsigned int i=0; i<targets->size(); i++)
+    for(size_t i=0; i<targets->size(); i++)
       {
       if (name == (*targets)[i]->Target->GetName())
         {
@@ -483,7 +483,7 @@ cmExportInstallFileGenerator
       {
       std::vector<cmInstallExportGenerator const*> const* installs =
                                                  exportSet->GetInstallations();
-      for(unsigned int i=0; i<installs->size(); i++)
+      for(size_t i=0; i<installs->size(); i++)
         {
         namespaces.push_back((*installs)[i]->GetNamespace());
         }
